% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{density.ngvb}
\alias{density.ngvb}
\title{Posterior density of a component's non-Gaussianity parameter eta.}
\usage{
\method{density}{ngvb}(x, component = x$comp.names[1L], ...)
}
\arguments{
\item{x}{An `ngvb` object.}

\item{component}{Component name (default: the first).}

\item{...}{Passed to [stats::density()].}
}
\value{
A `density` object (see [stats::density()]); `plot()` it directly.
}
\description{
For `method = "SVI"` this is a kernel density estimate over the GIG
variational factor's Monte-Carlo sample; for `"SCVI"` it is the weighted
quadrature grid underlying `scvi_update()`. Either way it is `q(eta)`, the
actual (approximate) posterior -- not just its mean.
}
