% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.ngvb}
\alias{fitted.ngvb}
\title{Fitted values of the underlying INLA fit.}
\usage{
\method{fitted}{ngvb}(object, ...)
}
\arguments{
\item{object}{An `ngvb` object.}

\item{...}{Ignored.}
}
\description{
Fitted values of the underlying INLA fit.
}
