% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jumpts}
\alias{jumpts}
\title{Simulated time series with two jumps}
\format{
A data frame with 100 rows and 2 columns:
\describe{
  \item{x}{integer time index, 1..100}
  \item{y}{observed value}
}
}
\usage{
jumpts
}
\description{
A 100-point series used to illustrate a non-Gaussian RW1 latent process. The
underlying signal is smooth except for two sudden jumps, which a Gaussian RW1
over-smooths and a non-Gaussian RW1 accommodates.
}
\examples{
plot(jumpts)
}
\keyword{datasets}
