% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\name{mungeGibbs.mixing,ngvb.list-method}
\alias{mungeGibbs.mixing,ngvb.list-method}
\title{Process \code{ngvb.list} when \code{method = "Gibbs"}. Produce a matrix with the
samples of the mixing variables \eqn{\mathbf{V}}.}
\usage{
\S4method{mungeGibbs.mixing}{ngvb.list}(object)
}
\arguments{
\item{object}{An ngvb.list object (output of \code{ngvb} function)}
}
\description{
Process \code{ngvb.list} when \code{method = "Gibbs"}. Produce a matrix with the
samples of the mixing variables \eqn{\mathbf{V}}.
}
