% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\name{mungeGibbs.ng,ngvb.list-method}
\alias{mungeGibbs.ng,ngvb.list-method}
\title{Process \code{ngvb.list} when \code{method = "Gibbs"}. Produce a matrix with the
samples of the non-Gaussianity parameter \eqn{\eta}.}
\usage{
\S4method{mungeGibbs.ng}{ngvb.list}(object)
}
\arguments{
\item{object}{An ngvb.list object (output of \code{ngvb} function)}
}
\description{
Process \code{ngvb.list} when \code{method = "Gibbs"}. Produce a matrix with the
samples of the non-Gaussianity parameter \eqn{\eta}.
}
