% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_check.R
\name{ng_sens_fixed}
\alias{ng_sens_fixed}
\title{Sensitivity of fixed effects to the non-Gaussianity parameter (Theorem 2).}
\usage{
ng_sens_fixed(b, gii, s12, u1, h)
}
\description{
Sensitivity of fixed effects to the non-Gaussianity parameter (Theorem 2).
}
\keyword{internal}
