% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\docType{class}
\name{ngvb.list-class}
\alias{ngvb.list-class}
\alias{ngvb.list}
\title{This class stores the output of the ngvb function.}
\description{
This class stores the output of the ngvb function.
}
\section{Slots}{

\describe{
\item{\code{history}}{List containing summaries of the model at each iteration.}

\item{\code{LGM}}{inla object containing summaries and marginals of the latent field \eqn{\mathbf{x}}
and hyperparameters \eqn{\boldsymbol{\theta}} of the last iteration.}

\item{\code{summary.mixing}}{Data frame containing summaries of the mixing variables \eqn{\mathbf{V}} (last iteration).}

\item{\code{summary.ng}}{Data frame containing summaries of the non-Gaussianiy parameters \eqn{\eta} (last iteration).}

\item{\code{configs}}{List containing the LnGM model specifications.}
}}

