% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{ngvb_detect_operator}
\alias{ngvb_detect_operator}
\title{Detect (or accept) the ngvb operator for one fitted component.}
\usage{
ngvb_detect_operator(fit, comp.name, user.op = NULL)
}
\description{
Detect (or accept) the ngvb operator for one fitted component.
}
\keyword{internal}
