% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{ngvb_find_f_arg}
\alias{ngvb_find_f_arg}
\title{Find and evaluate a named argument of the f(<comp>, ...) term in a formula.
Used to recover, e.g., `season.length` or `Cmatrix` from the fitted call.}
\usage{
ngvb_find_f_arg(fit, comp.name, argname)
}
\description{
Find and evaluate a named argument of the f(<comp>, ...) term in a formula.
Used to recover, e.g., `season.length` or `Cmatrix` from the fitted call.
}
\keyword{internal}
