% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{ngvb_find_f_model}
\alias{ngvb_find_f_model}
\title{Find the `model =` expression of the f(<comp>, ...) term in a formula.}
\usage{
ngvb_find_f_model(formula, comp.name)
}
\description{
Find the `model =` expression of the f(<comp>, ...) term in a formula.
}
\keyword{internal}
