% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ngvb_make_fit_V}
\alias{ngvb_make_fit_V}
\title{Build the conditional-on-V INLA fitter for a set of components.}
\usage{
ngvb_make_fit_V(fit, ops, comp.names)
}
\description{
Build the conditional-on-V INLA fitter for a set of components.
}
\keyword{internal}
