% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{ngvb_theme}
\alias{ngvb_theme}
\title{A shared, recessive ggplot theme for the ngvb plots.}
\usage{
ngvb_theme(base = 12)
}
\description{
A shared, recessive ggplot theme for the ngvb plots.
}
\keyword{internal}
