% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Orthodont}
\alias{Orthodont}
\title{Orthodontic growth data}
\format{
A data frame with 108 rows and 6 columns:
\describe{
  \item{subject}{subject id (random-intercept index)}
  \item{subject2}{subject id (random-slope index; distinct name required by INLA)}
  \item{Female}{indicator for female subjects}
  \item{time}{age recentred to 0,1,2,3}
  \item{tF}{Female x time interaction}
  \item{value}{measured distance}
}
}
\source{
Derived from the \code{Orthodont} data of Potthoff and Roy (1964),
  as distributed with the \pkg{nlme} package.
}
\usage{
Orthodont
}
\description{
Growth measurements on 27 children (16 boys, 11 girls) at ages 8, 10, 12, 14,
preprocessed for a linear mixed model with random intercepts and slopes. Used
to illustrate a two-component (random intercept + random slope) latent
non-Gaussian model.
}
\examples{
summary(Orthodont)
}
\keyword{datasets}
