% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{plot.gaussdiag}
\alias{plot.gaussdiag}
\title{Print method for "gaussdiag" class}
\usage{
plot.gaussdiag(x, n = rep(0, nrow(x$BF.check)), samples = NULL, ...)
}
\arguments{
\item{x}{List. Output of the \code{ng.check} function}

\item{n}{Integer. Plots the index number of the \code{n} indeces with the largest BF sensitivity.}

\item{samples}{Optional vector. Samples of the reference distribution.}
}
\value{
A list of ggplots:
\itemize{
   \item \code{First:}  The contribution of each index to the overallBF sensitivity: \eqn{d_i(\mathbf{y})}.
   \item \code{Second:} Reference distribution of the BF sensitivity vs observed BF sensitivity.
 }
}
\description{
Plots diagnostics that from the output of the \code{ng.check} function.
It includes the contribution of each index in the latent field to the overall Bayes factor (BF)
and the obsedved BF sensitivity compared against the reference distribution.
}
