% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ngvb}
\alias{plot.ngvb}
\title{Diagnostic plots for an ngvb fit: mixing variables per component and the
evolution of the non-Gaussianity parameters.}
\usage{
\method{plot}{ngvb}(x, ...)
}
\arguments{
\item{x}{An `ngvb` object.}

\item{...}{Passed to `plot`.}
}
\description{
Diagnostic plots for an ngvb fit: mixing variables per component and the
evolution of the non-Gaussianity parameters.
}
