% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\name{plot,ngvb.list,missing-method}
\alias{plot,ngvb.list,missing-method}
\title{Plots several summaries of an \code{ngvb.list} object.}
\usage{
\S4method{plot}{ngvb.list,missing}(x, y, ...)
}
\arguments{
\item{x}{An ngvb.list object (output of \code{ngvb} function)}

\item{y}{Not used.}

\item{...}{Extra arguments to be used in \code{plot(x@LGM, ...)} where \code{x@LGM} is an inla object.}
}
\description{
Plots several summaries of an \code{ngvb.list} object.
}
