% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\name{summary,ngvb.list-method}
\alias{summary,ngvb.list-method}
\title{Produces summaries of the \code{ngvb.list} object.}
\usage{
\S4method{summary}{ngvb.list}(object)
}
\arguments{
\item{object}{An ngvb.list object (output of \code{ngvb} function)}
}
\description{
Produces summaries of the \code{ngvb.list} object.
}
