% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{areal.plot}
\alias{areal.plot}
\title{Plot areal data}
\usage{
areal.plot(map, data, palette = "YlOrRd", title = "Value")
}
\arguments{
\item{data}{Vector of data to plot.}

\item{palette}{Character. Palette to use in Leaflet widget.}

\item{title}{Character. Title for the legend.}

\item{SpatialPolygons}{Spatial Polygons object.}
}
\value{
Leaflet widget
}
\description{
Generates a Leaflet widget to plot areal data. Requires \code{leaflet} and
\code{leaflet.extra} package.
}
