% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\name{fitted,ngvb.list-method}
\alias{fitted,ngvb.list-method}
\title{Summary of the fitted values. Also returns the marginals of the fitted values
if \code{compute=TRUE} in \code{control.predictor} of the inla object \code{fit}.}
\usage{
\S4method{fitted}{ngvb.list}(object)
}
\arguments{
\item{object}{An ngvb.list object (output of \code{ngvb} function)}
}
\description{
Summary of the fitted values. Also returns the marginals of the fitted values
if \code{compute=TRUE} in \code{control.predictor} of the inla object \code{fit}.
}
