% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{geo.plot}
\alias{geo.plot}
\title{Plot geostat data}
\usage{
geo.plot(
  data,
  mesh = NULL,
  n = 0,
  palette = c("blue", "white", "red"),
  domain = data[, 3],
  tile.provider = providers$Stamen.Terrain
)
}
\arguments{
\item{data}{Data frame. Should countain 3 collumns with the latitude, longitude and measurements.}

\item{mesh}{inla.mesh object.}

\item{n}{Integer. Plots the index number of the \code{n} indeces with the largest score.}

\item{palette}{Character. Palette to use in Leaflet widget.}

\item{domain}{Vector. Domain to use in palette scale}

\item{tile.provider}{Character. Name of the Leaflet tile provider.}
}
\value{
Leaflet widget
}
\description{
Generates a Leaflet widget to plot geostat data. Requires \code{leaflet} and
\code{leaflet.extra} package.
}
