% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jumpts}
\alias{jumpts}
\title{Time series with 2 jumps}
\format{
An object of class \code{data.frame}.
}
\usage{
jumpts
}
\description{
Time series of length 100 with two distinct jumps.
}
\keyword{datasets}
