% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\name{mungeGibbs.ng,ngvb.list-method}
\alias{mungeGibbs.ng,ngvb.list-method}
\title{Process \code{ngvb.list} when \code{method = "Gibbs"}. Produces a matrix with the
samples of the non-Gaussianity parameter \eqn{\eta}.}
\usage{
\S4method{mungeGibbs.ng}{ngvb.list}(object)
}
\arguments{
\item{object}{An ngvb.list object (output of \code{ngvb} function)}
}
\description{
Process \code{ngvb.list} when \code{method = "Gibbs"}. Produces a matrix with the
samples of the non-Gaussianity parameter \eqn{\eta}.
}
\examples{
\dontrun{
 #Here we fit an RW1 latent process to the jumpts time series
 plot(jumpts)

 #Fit LGM with INLA
 LGM     <- inla(y ~ -1 + f(x,  model = "rw1"),
                 data = jumpts)

 #Run 10 iterations from the Gibbs sampler
 LnGM.Gibbs <- ngvb(fit = LGM, selection = list(x=1:100),
                    method = "Gibbs", iter=10)

 Gibbs.V    <- mungeGibbs.mixing(LnGM.Gibbs)
 Gibbs.eta  <- mungeGibbs.ng(LnGM.Gibbs)
 }
}
