% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{ngvb_operator}
\alias{ngvb_operator}
\title{Construct an ngvb operator descriptor.}
\usage{
ngvb_operator(type, ...)
}
\arguments{
\item{type}{Model type, e.g. "ar1".}

\item{...}{Model-specific arguments (see the `op_*` builders).}
}
\value{
An operator descriptor list.
}
\description{
Construct an ngvb operator descriptor.
}
