% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ngvb_vb}
\alias{ngvb_vb}
\title{Multi-component structured variational inference loop.}
\usage{
ngvb_vb(
  inla.fit.V,
  ops,
  comp.names,
  method = c("SVI", "SCVI"),
  alpha.eta = 1,
  iter = 10,
  stop.rel.change = 0.001,
  n.sampling = 2000,
  verbose = TRUE
)
}
\description{
Multi-component structured variational inference loop.
}
\keyword{internal}
