% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Orthodont}
\alias{Orthodont}
\title{Data of an orthodontic study.}
\format{
An object of class \code{data.frame}.
}
\usage{
Orthodont
}
\description{
Original dataset \code{potthoffroy} from package \code{mice}.
Modified to be used by INLA.
}
\keyword{datasets}
