% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\name{plot,ngvb.list,missing-method}
\alias{plot,ngvb.list,missing-method}
\title{Plots several summaries of an \code{ngvb.list} object.}
\usage{
\S4method{plot}{ngvb.list,missing}(x, y, ...)
}
\arguments{
\item{x}{An ngvb.list object (output of \code{ngvb} function)}

\item{y}{Not used.}

\item{...}{Extra arguments to be used in \code{plot(x@LGM, ...)} where \code{x@LGM} is an inla object.}
}
\description{
Plots several summaries of an \code{ngvb.list} object.
}
\examples{
\dontrun{
 #Here we fit an RW1 latent process to the jumpts time series
 plot(jumpts)

 #Fit LGM with INLA
 LGM     <- inla(y ~ -1 + f(x,  model = "rw1"),
                 data = jumpts)

 #Fit LnGM with ngvb
 LnGM <- ngvb(fit = LGM, selection = list(x=1:100))

 #Available methods
 summary(LnGM)
 print(LnGM)
 plot(LnGM)
 fitted(LnGM)
 samples <- simulate(LnGM)
 }
}
