% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngvb.list.R
\name{print,ngvb.list-method}
\alias{print,ngvb.list-method}
\title{Prints outputs of the \code{ngvb.list} object.}
\usage{
\S4method{print}{ngvb.list}(x)
}
\arguments{
\item{x}{An ngvb.list object (output of \code{ngvb} function)}
}
\description{
Prints outputs of the \code{ngvb.list} object.
}
